slideshow_app/requirements.txt
attrs==23.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cryptography==42.0.3
cycler==0.12.1
decorator==4.4.2
distlib==0.3.8
filelock==3.13.1
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
fonttools==4.49.0
greenlet==3.0.3
h11==0.14.0
html5lib==1.1
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
moviepy==1.0.3
mysql-connector-python==8.3.0
numpy==1.26.4
opencv-python==4.9.0.80
outcome==1.3.0.post0
packaging==23.2
pillow==10.2.0
platformdirs==4.2.0
proglog==0.1.10
psycopg2-binary
pycparser==2.21
PyJWT==2.1.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
requests==2.31.0
scipy==1.12.0
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.27
tqdm==4.66.2
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.1.0
virtualenv==20.25.0
webencodings==0.5.1
Werkzeug==3.0.1
wordcloud==1.9.3
wsproto==1.2.0
